Fully working registration and login
